requirements.txt
flask
requests
python-dotenv
playwright
pytest
pytest-playwright
pytest-html
allure-pytest
slack-sdk
